Extract dataset code in to its own component
